pyproject.toml
[project]
name = "black-flames"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.6.3",
    "jishaku>=2.6.0",
    "motor>=3.7.1",
    "python-dotenv>=1.1.1",
    "umongo==3.1.0",
    "marshmallow==3.21.3",
    "pychallonge>=1.11.5",
]
